Allow shipments to be created independently of labels.
