+++ lib/fedex/request/shipment.rb
@@ -1,0 +1,106 @@
+require 'fedex/request/base'
+
+module Fedex
+  module Request
+    class Shipment < Base
+      attr_reader :response_details
+
+      def initialize(credentials, options={})
+        super
+        requires! options
+        # Label specification is required even if we're not using it.
+        @label_specification = {
+          :label_format_type => 'COMMON2D',
+          :image_type => 'PDF',
+          :label_stock_type => 'PAPER_LETTER'
+        }
+        @label_specification.merge! options[:label_specification] if options[:label_specification]
+      end
+
+      # Sends post request to Fedex web service and parse the response.
+      # A label file is created with the label at the specified location.
+      # The parsed Fedex response is available in #response_details
+      # e.g. response_details[:completed_shipment_detail][:completed_package_details][:tracking_ids][:tracking_number]
+      def process_request
+        api_response = self.class.post api_url, :body => build_xml
+        puts api_response if @debug
+        response = parse_response(api_response)
+        if success?(response)
+          success_response(api_response, response)
+        else
+          failure_response(api_response, response)
+        end
+      end
+
+      private
+
+      # Add information for shipments
+      def add_requested_shipment(xml)
+        xml.RequestedShipment{
+          xml.ShipTimestamp Time.now.utc.iso8601(2)
+          xml.DropoffType @shipping_options[:drop_off_type] ||= "REGULAR_PICKUP"
+          xml.ServiceType service_type
+          xml.PackagingType @shipping_options[:packaging_type] ||= "YOUR_PACKAGING"
+          add_shipper(xml)
+          add_recipient(xml)
+          add_shipping_charges_payment(xml)
+          add_customs_clearance(xml) if @customs_clearance
+          add_custom_components(xml)
+          xml.RateRequestTypes "ACCOUNT"
+          add_packages(xml)
+        }
+      end
+
+      # Hook that can be used to add custom parts.
+      def add_custom_components(xml)
+        add_label_specification xml
+      end
+
+      # Add the label specification
+      def add_label_specification(xml)
+        xml.LabelSpecification {
+          hash_to_xml(xml, @label_specification)
+        }
+      end
+
+      # Callback used after a failed shipment response.
+      def failure_response(api_response, response)
+        error_message = if response[:process_shipment_reply]
+          [response[:process_shipment_reply][:notifications]].flatten.first[:message]
+        else
+          api_response["Fault"]["detail"]["fault"]["reason"]
+        end rescue $1
+        raise RateError, error_message
+      end
+
+      # Callback used after a successful shipment response.
+      def success_response(api_response, response)
+        @response_details = response[:process_shipment_reply]
+      end
+
+      # Build xml Fedex Web Service request
+      def build_xml
+        builder = Nokogiri::XML::Builder.new do |xml|
+          xml.ProcessShipmentRequest(:xmlns => "http://fedex.com/ws/ship/v10"){
+            add_web_authentication_detail(xml)
+            add_client_detail(xml)
+            add_version(xml)
+            add_requested_shipment(xml)
+          }
+        end
+        builder.doc.root.to_xml
+      end
+
+      def service_id
+        'ship'
+      end
+
+      # Successful request
+      def success?(response)
+        response[:process_shipment_reply] &&
+          %w{SUCCESS WARNING NOTE}.include?(response[:process_shipment_reply][:highest_severity])
+      end
+
+    end
+  end
+end
