+++ lib/fedex/shipment.rb
@@ -36,5 +36,13 @@ module Fedex
       Request::Rate.new(@credentials, options).process_request
     end
 
+    # @param [Hash] shipper, A hash containing the shipper information
+    # @param [Hash] recipient, A hash containing the recipient information
+    # @param [Array] packages, An arrary including a hash for each package being shipped
+    # @param [String] service_type, A valid fedex service type, to view a complete list of services Fedex::Shipment::SERVICE_TYPES
+    def ship(options = {})
+      Request::Shipment.new(@credentials, options).process_request
+    end
+
   end
 end
